+++ ProductivityBooster/src/application/Main.java
@@ -1,4 +1,4 @@
- package application;
+package application;
 	
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
