+++ ProductivityBooster/src/application/main.fxml
@@ -13,6 +13,7 @@
 <?import javafx.scene.text.Text?>
 
 <BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.MainController">
+   <!-- <fx:include fx:id="newTask" source="newTask.fxml"/> -->
    <left>
       <VBox prefHeight="400.0" prefWidth="131.0" style="-fx-background-color: #233D4D;" BorderPane.alignment="CENTER">
          <children>
